+++ Source/Enemies/Enemy.c
@@ -19,7 +19,7 @@ extern	OGLPoint3D			gCoord;
 extern	OGLVector3D			gDelta;
 extern	short				gNumCollisions;
 extern	CollisionRec		gCollisionList[];
-extern	u_long		gAutoFadeStatusBits;
+extern	uint32_t		gAutoFadeStatusBits;
 extern	OGLVector3D		gRecentTerrainNormal;
 
 
@@ -227,11 +227,11 @@ float	x,z;
 // OUTPUT: true = was deleted
 //
 
-Boolean DoEnemyCollisionDetect(ObjNode *theEnemy, u_long ctype, Boolean useBBoxBottom)
+Boolean DoEnemyCollisionDetect(ObjNode *theEnemy, uint32_t ctype, Boolean useBBoxBottom)
 {
 float	terrainY,distToFloor,bottomOff;
 int		i;
-u_long	wasInWater;
+uint32_t	wasInWater;
 
 
 	wasInWater = (theEnemy->StatusBits & STATUS_BIT_UNDERWATER);		// see if was in H2O on previous frame
