+++ Source/Headers/globals.h
@@ -34,7 +34,7 @@
 
 #define Alloc_2d_array(type, array, n, m)								\
 {																		\
-u_long _i;																\
+uint32_t _i;																\
 																		\
 	array = (type **) AllocPtr((long)(n) * sizeof(type *));				\
 	if (array == nil)													\
@@ -42,7 +42,7 @@ u_long _i;																\
 	array[0] = (type *) AllocPtr((long)(n) * (long)(m) * sizeof(type));	\
 	if (array[0] == nil)												\
 		DoFatalAlert("Alloc_2d_array failed!");						\
-	for (_i = 1; _i < (u_long)(n); _i++)										\
+	for (_i = 1; _i < (uint32_t)(n); _i++)										\
 		array[_i] = array[_i-1] + (m);									\
 }
